requirements.txt
python-dotenv
requests
google-cloud-bigquery